9#피드백반영

+++ src/hooks/useComment.ts
@@ -47,21 +47,29 @@ const useComment = (eventId: number, initComments?: Comment[]) => {
       }
     },
     remove: async (commentId: number) => {
-      const rs = await deleteComment(commentId);
-      if (rs) {
-        responseHandler(rs, {});
+      if (confirm("정말 댓글을 삭제 하시겠습니까?")) {
+        const rs = await deleteComment(commentId);
+        if (rs) {
+          responseHandler(rs, {});
+        }
+      } else {
+        return;
       }
     },
     modify: async (commentId: number, content: string) => {
-      const rs = await patchComment(content, commentId);
+      if (confirm("댓글을 수정 하시겠습니까?")) {
+        const rs = await patchComment(content, commentId);
 
-      if (rs) {
-        const handler = {
-          success: () => {
-            setIsModify({ status: false, commentId: -1 });
-          },
-        };
-        responseHandler(rs, handler);
+        if (rs) {
+          const handler = {
+            success: () => {
+              setIsModify({ status: false, commentId: -1 });
+            },
+          };
+          responseHandler(rs, handler);
+        }
+      } else {
+        return;
       }
     },
   };
